shared/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/shared"]

[project]
name = "dark-factory-shared"
version = "0.1.0"
description = "Shared utilities for the dark-factory software factory"
requires-python = ">=3.11,<4"
dependencies = ["pydantic>=2.7", "aiosqlite>=0.20", "pyyaml>=6.0", "python-dotenv>=1.0"]

[project.optional-dependencies]
vllm = ["aiohttp>=3.9"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=7.1.0",
    "ruff>=0.4",
    "aiohttp>=3.9",
]

[tool.pytest.ini_options]
markers = ["integration: live CLI tests requiring OAuth accounts (deselect: -m 'not integration')"]
addopts = "-m 'not integration'"
asyncio_mode = "auto"

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "UP", "B", "SIM", "I"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
